Move session-related utils to their own file
